Changes made:
Removed .css file from application package along with the associated
line in Main that invoked it.
Merged the two Main classes, kept the one in application, deleted the
start package.
Added a class description to Main.
