+++ GroupProject2/src/buttons/AwayButton.java
@@ -9,7 +9,7 @@ import states.AlarmSystemContext;
  * The AwayButton class is for when the Away button of the GUI is clicked, as
  * part of the AlarmSystem's button hierarchy.
  * 
- * @author Ethan Nunn, Brian Le, Colin Bolduc, Danield Renaud and Zachary
+ * @author Ethan Nunn, Brian Le, Colin Bolduc, Daniel Renaud and Zachary
  *         Boling-Green
  */
 public class AwayButton extends GUIButton implements EventHandler<ActionEvent> {
