+++ GroupProject2/src/buttons/CancelButton.java
@@ -9,11 +9,10 @@ import states.AlarmSystemContext;
  * The CancelButton class is for when the Cancel button of the GUI is clicked,
  * as part of the AlarmSystem's button hierarchy.
  * 
- * @author Ethan Nunn, Brian Le, Colin Bolduc, Danield Renaud and Zachary
+ * @author Ethan Nunn, Brian Le, Colin Bolduc, Daniel Renaud and Zachary
  *         Boling-Green
  */
-public class CancelButton extends GUIButton
-		implements EventHandler<ActionEvent> {
+public class CancelButton extends GUIButton implements EventHandler<ActionEvent> {
 
 	/**
 	 * Create the button with the proper display
