+++ GroupProject2/src/buttons/GUICheckBox.java
@@ -8,11 +8,10 @@ import javafx.scene.control.CheckBox;
  * The abstract GUI CheckBox object. Helps to get rid of conditionals and ease
  * of future modification of the system.
  *
- * @author Ethan Nunn, Brian Le, Colin Bolduc, Danield Renaud and Zachary
+ * @author Ethan Nunn, Brian Le, Colin Bolduc, Daniel Renaud and Zachary
  *         Boling-Green
  */
-public abstract class GUICheckBox extends CheckBox
-		implements EventHandler<ActionEvent> {
+public abstract class GUICheckBox extends CheckBox implements EventHandler<ActionEvent> {
 	/**
 	 * Create the checkbox with the proper text. Makes the button a listener to
 	 * clicks on itself.
