+++ GroupProject2/src/buttons/MotionDetectorButton.java
@@ -10,11 +10,10 @@ import states.AlarmSystemContext;
  * GUI is clicked (simulating a motion sensor), as part of the AlarmSystem's
  * button hierarchy.
  * 
- * @author Ethan Nunn, Brian Le, Colin Bolduc, Danield Renaud and Zachary
+ * @author Ethan Nunn, Brian Le, Colin Bolduc, Daniel Renaud and Zachary
  *         Boling-Green
  */
-public class MotionDetectorButton extends GUIButton
-		implements EventHandler<ActionEvent> {
+public class MotionDetectorButton extends GUIButton implements EventHandler<ActionEvent> {
 
 	/**
 	 * Create the button with the proper display
@@ -31,7 +30,6 @@ public class MotionDetectorButton extends GUIButton
 	 */
 	@Override
 	public void handle(ActionEvent event) {
-		AlarmSystemContext.instance()
-				.handleEvent(MotionDetectionEvent.instance());
+		AlarmSystemContext.instance().handleEvent(MotionDetectionEvent.instance());
 	}
 }
