+++ GroupProject2/src/buttons/StayButton.java
@@ -9,7 +9,7 @@ import states.AlarmSystemContext;
  * The StayButton class is for when the Stay button of the GUI is clicked, as
  * part of the AlarmSystem's button hierarchy.
  * 
- * @author Ethan Nunn, Brian Le, Colin Bolduc, Danield Renaud and Zachary
+ * @author Ethan Nunn, Brian Le, Colin Bolduc, Daniel Renaud and Zachary
  *         Boling-Green
  */
 public class StayButton extends GUIButton implements EventHandler<ActionEvent> {
